% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AbandonedCheckout.R
\name{AbandonedCheckout}
\alias{AbandonedCheckout}
\alias{getAbandonedCheckouts}
\alias{getAbandonedCheckoutsCount}
\title{AbandonedCheckout Functions}
\usage{
getAbandonedCheckouts(...)

getAbandonedCheckoutsCount(...)
}
\arguments{
\item{...}{additional request parameters}
}
\value{
a list containing a AbandonedCheckout fields or a list of AbandonedCheckouts
}
\description{
AbandonedCheckout Functions
}
\references{
Shopify API AbandonedCheckout documentation: \url{https://help.shopify.com/api/reference/orders/abandoned_checkouts}
}
