% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AccessScope.R
\name{AccessScope}
\alias{AccessScope}
\alias{getAccessScopes}
\title{AccessScope Functions}
\usage{
getAccessScopes(...)
}
\arguments{
\item{...}{additional request parameters}
}
\value{
a list containing an AccessScope fields or a list of AccessScopes
}
\description{
AccessScope Functions
}
\references{
Shopify API AccessScope documentation: \url{https://help.shopify.com/api/reference/access/accessscope}
}
