% Generated by roxygen2 (4.0.1): do not edit by hand
\name{Announcements}
\alias{Announcements}
\alias{showAnnouncements}
\title{Announcement Functions}
\usage{
showAnnouncements(current = TRUE)
}
\arguments{
\item{current}{show only non-expired announcements (default \code{TRUE})}
}
\value{
invisibly returns the \code{\link{ShopifyShop}} object to allow for chaining
}
\description{
Announcement Functions
}

