% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ApplicationCharge.R
\name{ApplicationCharge}
\alias{ApplicationCharge}
\alias{createApplicationCharge}
\alias{getApplicationCharge}
\alias{getApplicationCharges}
\alias{activateApplicationCharge}
\title{ApplicationCharge Functions}
\usage{
createApplicationCharge(charge, ...)

getApplicationCharge(chargeId, ...)

getApplicationCharges(...)

activateApplicationCharge(charge, ...)
}
\arguments{
\item{charge}{a list containing ApplicationCharge fields}

\item{...}{additional request parameters}

\item{chargeId}{an ApplicationCharge id number}
}
\value{
a list containing an ApplicationCharge fields or a list of ApplicationCharges
}
\description{
ApplicationCharge Functions
}
\references{
Shopify API ApplicationCharge documentation: \url{https://help.shopify.com/api/reference/billing/applicationcharge}
}
