% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ApplicationCredit.R
\name{ApplicationCredit}
\alias{ApplicationCredit}
\alias{createApplicationCredit}
\alias{getApplicationCredit}
\alias{getApplicationCredits}
\title{ApplicationCredit Functions}
\usage{
createApplicationCredit(credit, ...)

getApplicationCredit(creditId, ...)

getApplicationCredits(...)
}
\arguments{
\item{credit}{a list containing ApplicationCredit fields}

\item{...}{additional request parameters}

\item{creditId}{an ApplicationCredit id number}
}
\value{
a list containing an ApplicationCredit fields or a list of ApplicationCredits
}
\description{
ApplicationCredit Functions
}
\references{
Shopify API ApplicationCredit documentation: \url{https://help.shopify.com/api/reference/billing/applicationcredit}
}
