% Generated by roxygen2 (4.0.1): do not edit by hand
\name{Article}
\alias{Article}
\alias{createArticle}
\alias{deleteArticle}
\alias{getArticle}
\alias{getArticleAuthors}
\alias{getArticleTags}
\alias{getArticles}
\alias{getArticlesCount}
\alias{modifyArticle}
\title{Article Functions}
\usage{
getArticles(blogId, ...)

getArticlesCount(blogId, ...)

getArticle(blogId, articleId, ...)

createArticle(blogId, article, ...)

modifyArticle(blogId, article, ...)

getArticleAuthors(...)

getArticleTags(...)

deleteArticle(blogId, articleId, ...)
}
\arguments{
\item{blogId}{a Blog id number}

\item{articleId}{an Article id number}

\item{article}{a list containing Article fields}

\item{...}{additional request parameters}
}
\value{
a list containing an Article fields or a list of Articles
}
\description{
Article Functions
}
\references{
Shopify API Article documentation: \url{http://docs.shopify.com/api/article}
}

