% Generated by roxygen2 (4.0.1): do not edit by hand
\name{Asset}
\alias{Asset}
\alias{createAsset}
\alias{deleteAsset}
\alias{getAsset}
\alias{getAssets}
\alias{modifyAsset}
\title{Asset Functions}
\usage{
getAssets(themeId, ...)

getAsset(themeId, assetKey, ...)

createAsset(themeId, asset, ...)

deleteAsset(themeId, assetKey, ...)
}
\arguments{
\item{themeId}{a Theme id number}

\item{assetKey}{an Asset key e.g. \code{"templates/index.liquid"}}

\item{asset}{a  list containing Asset fields}

\item{...}{additional request parameters}
}
\value{
a list containing an Asset fields or a list of Assets
}
\description{
Asset Functions
}
\references{
Shopify API Asset documentation: \url{http://docs.shopify.com/api/asset}
}

