% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Blog.R
\name{Blog}
\alias{Blog}
\alias{getBlogs}
\alias{getBlogsCount}
\alias{getBlog}
\alias{createBlog}
\alias{modifyBlog}
\alias{deleteBlog}
\title{Blog Functions}
\usage{
getBlogs(...)

getBlogsCount(...)

getBlog(blogId, ...)

createBlog(blog, ...)

modifyBlog(blog, ...)

deleteBlog(blogId, ...)
}
\arguments{
\item{...}{additional request parameters}

\item{blogId}{a Blog id number}

\item{blog}{a list containing Blog fields}
}
\value{
a list containing a Blog fields or a list of Blogs
}
\description{
Blog Functions
}
\references{
Shopify API Blog documentation: \url{https://help.shopify.com/api/reference/online_store/blog}
}
