% Generated by roxygen2 (4.0.1): do not edit by hand
\name{CarrierService}
\alias{CarrierService}
\alias{createCarrierService}
\alias{deleteCarrierService}
\alias{getCarrierService}
\alias{getCarrierServices}
\alias{modifyCarrierService}
\title{CarrierService Functions}
\usage{
createCarrierService(service, ...)

modifyCarrierService(service, ...)

getCarrierServices(...)

getCarrierService(serviceId, ...)

deleteCarrierService(serviceId, ...)
}
\arguments{
\item{serviceId}{a CarrierService id number}

\item{service}{a list containing CarrierService fields}

\item{...}{additional request parameters}
}
\value{
a list containing a CarrierService fields or a list of CarrierServices
}
\description{
CarrierService Functions
}
\references{
Shopify API CarrierService documentation: \url{http://docs.shopify.com/api/carrierservice}
}

