% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Checkout.R
\name{Checkout}
\alias{Checkout}
\alias{getCheckoutsCount}
\alias{getCheckouts}
\title{Checkout Functions}
\usage{
getCheckoutsCount(...)

getCheckouts(...)
}
\arguments{
\item{...}{additional request parameters}
}
\value{
a list of Checkouts
}
\description{
Checkout Functions
}
\references{
Shopify API Checkout documentation: \url{https://help.shopify.com/api/reference/sales_channels/checkout}
}
