% Generated by roxygen2 (4.0.1): do not edit by hand
\name{Collect}
\alias{Collect}
\alias{createCollect}
\alias{deleteCollect}
\alias{getCollect}
\alias{getCollects}
\alias{getCollectsCount}
\title{Collect Functions}
\usage{
createCollect(collect, ...)

deleteCollect(collectId, ...)

getCollects(...)

getCollectsCount(...)

getCollect(collectId, ...)
}
\arguments{
\item{collectId}{a Collect id number}

\item{collect}{a list containing Collect fields}

\item{...}{additional request parameters}
}
\value{
a list containing a Collect fields or a list of Collects
}
\description{
Collect Functions
}
\references{
Shopify API Collect documentation: \url{http://docs.shopify.com/api/collect}
}

