% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CollectionListing.R
\name{CollectionListing}
\alias{CollectionListing}
\alias{createCollectionListing}
\alias{deleteCollectionListing}
\alias{getCollectionListings}
\alias{getCollectionListingCount}
\alias{getCollectionListingProductIds}
\title{CollectionListing Functions}
\usage{
createCollectionListing(collectionId, ...)

deleteCollectionListing(collectionListingId, ...)

getCollectionListings(...)

getCollectionListingCount(...)

getCollectionListingProductIds(collectionListingId, ...)
}
\arguments{
\item{collectionId}{a Collection id}

\item{...}{additional request parameters}

\item{collectionListingId}{a CollectionListing id}
}
\value{
a list of Checkouts
}
\description{
CollectionListing Functions
}
\references{
Shopify API CollectionListing documentation: \url{https://help.shopify.com/api/reference/sales_channels/checkout}
}
