% Generated by roxygen2 (4.0.1): do not edit by hand
\name{Comment}
\alias{Comment}
\alias{approveComment}
\alias{createComment}
\alias{getComment}
\alias{getComments}
\alias{getCommentsCount}
\alias{markCommentAsNotSpam}
\alias{markCommentAsSpam}
\alias{modifyComment}
\alias{removeComment}
\alias{restoreComment}
\title{Comment Functions}
\usage{
getComments(blogId, articleId, ...)

getCommentsCount(blogId, articleId, ...)

getComment(commentId, ...)

createComment(comment, ...)

modifyComment(comment, ...)

markCommentAsSpam(commentId, ...)

markCommentAsNotSpam(commentId, ...)

approveComment(commentId, ...)

removeComment(commentId, ...)

restoreComment(commentId, ...)
}
\arguments{
\item{blogId}{a Blog id number (leave blank to fetch all comments for the shop)}

\item{articleId}{an Article id number (leave blank to fetch all comments for this blog)}

\item{commentId}{a Comment id number}

\item{comment}{a list containing Comment fields}

\item{...}{additional request parameters}
}
\value{
a list containing a Comment fields or a list of Comments
}
\description{
Comment Functions
}
\references{
Shopify API Comment documentation: \url{http://docs.shopify.com/api/comment}
}

