% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Country.R
\name{Country}
\alias{Country}
\alias{getCountries}
\alias{getCountriesCount}
\alias{getCountry}
\alias{createCountry}
\alias{modifyCountry}
\alias{deleteCountry}
\title{Country Functions}
\usage{
getCountries(...)

getCountriesCount(...)

getCountry(countryId, ...)

createCountry(country, ...)

modifyCountry(country, ...)

deleteCountry(countryId, ...)
}
\arguments{
\item{...}{additional request parameters}

\item{countryId}{a Country id number}

\item{country}{a list containing Country fields}
}
\value{
a list containing a Country fields or a list of Countrys
}
\description{
Country Functions
}
\references{
Shopify API Country documentation: \url{https://help.shopify.com/api/reference/store_properties/country}
}
