% Generated by roxygen2 (4.0.1): do not edit by hand
\name{CustomCollection}
\alias{CustomCollection}
\alias{createCustomCollection}
\alias{deleteCustomCollection}
\alias{getCustomCollection}
\alias{getCustomCollections}
\alias{getCustomCollectionsCount}
\alias{modifyCustomCollection}
\title{CustomCollection Functions}
\usage{
getCustomCollections(...)

getCustomCollectionsCount(...)

getCustomCollection(customCollectionId, ...)

createCustomCollection(customCollection, ...)

modifyCustomCollection(customCollection, ...)

deleteCustomCollection(customCollectionId, ...)
}
\arguments{
\item{customCollectionId}{a CustomCollection id number}

\item{customCollection}{a list containing CustomCollection fields}

\item{...}{additional request parameters}
}
\value{
a list containing a CustomCollection fields or a list of CustomCollections
}
\description{
CustomCollection Functions
}
\references{
Shopify API CustomCollection documentation: \url{http://docs.shopify.com/api/custom_collection}
}

