% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Customer.R
\name{Customer}
\alias{Customer}
\alias{getCustomers}
\alias{searchCustomers}
\alias{getCustomer}
\alias{createCustomer}
\alias{modifyCustomer}
\alias{deleteCustomer}
\alias{getCustomersCount}
\alias{getCustomerOrders}
\title{Customer Functions}
\usage{
getCustomers(...)

searchCustomers(query, ...)

getCustomer(customerId, ...)

createCustomer(customer, ...)

modifyCustomer(customer, ...)

deleteCustomer(customerId, ...)

getCustomersCount(...)

getCustomerOrders(customerId, ...)
}
\arguments{
\item{...}{additional request parameters}

\item{query}{a query string to use when searching Customer records}

\item{customerId}{a Customer id number}

\item{customer}{a list containing Customer fields}
}
\value{
a list containing a Customer fields or a list of Customers
}
\description{
Customer Functions
}
\references{
Shopify API Customer documentation: \url{https://help.shopify.com/api/reference/customers/customer}
}
