% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CustomerAddress.R
\name{CustomerAddress}
\alias{CustomerAddress}
\alias{getCustomerAddresses}
\alias{getCustomerAddress}
\alias{createCustomerAddress}
\alias{modifyCustomerAddress}
\alias{deleteCustomerAddress}
\alias{setCustomerDefaultAddress}
\title{CustomerAddress Functions}
\usage{
getCustomerAddresses(customerId, ...)

getCustomerAddress(customerId, customerAddressId, ...)

createCustomerAddress(customerAddress, customerId, ...)

modifyCustomerAddress(customerAddress, customerId, ...)

deleteCustomerAddress(customerId, customerAddressId, ...)

setCustomerDefaultAddress(customerId, customerAddressId, ...)
}
\arguments{
\item{customerId}{a Customer id}

\item{...}{additional request parameters}

\item{customerAddressId}{a CustomerAddress id number}

\item{customerAddress}{a list containing CustomerAddress fields}
}
\value{
a list containing a CustomerAddress fields or a list of CustomerAddresss
}
\description{
CustomerAddress Functions
}
\references{
Shopify API CustomerAddress documentation: \url{https://help.shopify.com/api/reference/customers/customer_address}
}
