% Generated by roxygen2 (4.0.1): do not edit by hand
\name{CustomerGroup}
\alias{CustomerGroup}
\alias{CustomerSavedSearch}
\alias{createCustomerGroup}
\alias{deleteCustomerGroup}
\alias{getCustomerGroup}
\alias{getCustomerGroupCustomers}
\alias{getCustomerGroups}
\alias{getCustomerGroupsCount}
\alias{modifyCustomerGroup}
\title{CustomerGroup Functions}
\usage{
getCustomerGroups(...)

getCustomerGroupsCount(...)

getCustomerGroup(customerGroupId, ...)

getCustomerGroupCustomers(customerGroupId, ...)

createCustomerGroup(customerGroup, ...)

modifyCustomerGroup(customerGroup, ...)

deleteCustomerGroup(customerGroupId, ...)
}
\arguments{
\item{customerGroupId}{a CustomerGroup id number}

\item{customerGroup}{a list containing CustomerGroup fields}

\item{...}{additional request parameters}
}
\value{
a list containing a CustomerGroup fields or a list of CustomerGroups
}
\description{
CustomerGroup Functions
}
\references{
Shopify API CustomerGroup documentation: \url{http://docs.shopify.com/api/customergroup}
}

