% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CustomerSavedSearch.R
\name{CustomerSavedSearch}
\alias{CustomerSavedSearch}
\alias{getCustomerSavedSearches}
\alias{getCustomerSavedSearchesCount}
\alias{getCustomerSavedSearch}
\alias{getCustomerSavedSearchResults}
\alias{createCustomerSavedSearch}
\alias{modifyCustomerSavedSearch}
\alias{deleteCustomerSavedSearch}
\title{CustomerSavedSearch Functions}
\usage{
getCustomerSavedSearches(...)

getCustomerSavedSearchesCount(...)

getCustomerSavedSearch(customerSavedSearchId, ...)

getCustomerSavedSearchResults(customerSavedSearchId, ...)

createCustomerSavedSearch(customerSavedSearch, ...)

modifyCustomerSavedSearch(customerSavedSearch, ...)

deleteCustomerSavedSearch(customerSavedSearchId, ...)
}
\arguments{
\item{...}{additional request parameters}

\item{customerSavedSearchId}{a CustomerSavedSearch id number}

\item{customerSavedSearch}{a list containing CustomerSavedSearch fields}
}
\value{
a list containing a CustomerSavedSearch fields or a list of CustomerSavedSearchs
}
\description{
CustomerSavedSearch Functions
}
\references{
Shopify API CustomerSavedSearch documentation: \url{https://help.shopify.com/api/reference/customers/customersavedsearch}
}
