% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DraftOrder.R
\name{DraftOrder}
\alias{DraftOrder}
\alias{createDraftOrder}
\alias{modifyDraftOrder}
\alias{getDraftOrders}
\alias{getDraftOrder}
\alias{getDraftOrdersCount}
\alias{sendDraftOrderInvoice}
\alias{deleteDraftOrder}
\alias{completeDraftOrder}
\title{DraftOrder Functions}
\usage{
createDraftOrder(draftOrder, ...)

modifyDraftOrder(draftOrder, ...)

getDraftOrders(...)

getDraftOrder(draftOrderId, ...)

getDraftOrdersCount(...)

sendDraftOrderInvoice(draftOrderId, draftOrderInvoice, ...)

deleteDraftOrder(draftOrderId, ...)

completeDraftOrder(draftOrderId, ...)
}
\arguments{
\item{draftOrder}{a list containing DraftOrder fields}

\item{...}{additional request parameters}

\item{draftOrderId}{a DraftOrder id number}

\item{draftOrderInvoice}{an list of draft order invoice fields}
}
\value{
a list containing a DraftOrder fields or a list of DraftOrders
}
\description{
DraftOrder Functions
}
\references{
Shopify API DraftOrder documentation: \url{https://help.shopify.com/api/reference/orders/draft_order}
}
