% Generated by roxygen2 (4.0.1): do not edit by hand
\name{Event}
\alias{Event}
\alias{getEvent}
\alias{getEvents}
\alias{getEventsCount}
\title{Event Functions}
\usage{
getEvents(...)

getEvent(eventId, ...)

getEventsCount(...)
}
\arguments{
\item{eventId}{a Event id number}

\item{...}{additional request parameters}
}
\value{
a list containing a Event fields or a list of Events
}
\description{
Event Functions
}
\references{
Shopify API Event documentation: \url{http://docs.shopify.com/api/event}
}

