% Generated by roxygen2 (4.0.1): do not edit by hand
\name{Fulfillment}
\alias{Fulfillment}
\alias{cancelFulfillment}
\alias{completeFulfillment}
\alias{createFulfillment}
\alias{getFulfillment}
\alias{getFulfillments}
\alias{getFulfillmentsCount}
\alias{modifyFulfillment}
\title{Fulfillment Functions}
\usage{
getFulfillments(orderId, ...)

getFulfillmentsCount(orderId, ...)

getFulfillment(orderId, fulfillmentId, ...)

createFulfillment(orderId, fulfillment, ...)

modifyFulfillment(orderId, fulfillment, ...)

completeFulfillment(orderId, fulfillmentId, ...)

cancelFulfillment(orderId, fulfillmentId, ...)
}
\arguments{
\item{orderId}{an Order id number}

\item{fulfillmentId}{a Fulfillment id number}

\item{fulfillment}{a list containing Fulfillment fields}

\item{...}{additional request parameters}
}
\value{
a list containing a Fulfillment fields or a list of Fulfillments
}
\description{
Fulfillment Functions
}
\references{
Shopify API Fulfillment documentation: \url{http://docs.shopify.com/api/fulfillment}
}

