% Generated by roxygen2 (4.0.1): do not edit by hand
\name{FulfillmentService}
\alias{FulfillmentService}
\alias{createFulfillmentService}
\alias{deleteFulfillmentService}
\alias{getFulfillmentService}
\alias{getFulfillmentServices}
\alias{modifyFulfillmentService}
\title{FulfillmentService Functions}
\usage{
getFulfillmentServices(...)

createFulfillmentService(fulfillmentService, ...)

getFulfillmentService(fulfillmentServiceId, ...)

modifyFulfillmentService(fulfillmentService, ...)

deleteFulfillmentService(fulfillmentServiceId, ...)
}
\arguments{
\item{fulfillmentServiceId}{a FulfillmentService id number}

\item{fulfillmentService}{a list containing FulfillmentService fields}

\item{...}{additional request parameters}
}
\value{
a list containing a FulfillmentService fields or a list of FulfillmentServices
}
\description{
FulfillmentService Functions
}
\references{
Shopify API FulfillmentService documentation: \url{http://docs.shopify.com/api/fulfillmentservice}
}

