% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GiftCard.R
\name{GiftCard}
\alias{GiftCard}
\alias{getGiftCards}
\alias{getGiftCard}
\alias{getGiftCardsCount}
\alias{createGiftCard}
\alias{modifyGiftCard}
\alias{disableGiftCard}
\alias{searchGiftCards}
\title{GiftCard Functions}
\usage{
getGiftCards(...)

getGiftCard(giftCardId, ...)

getGiftCardsCount(...)

createGiftCard(giftCard, ...)

modifyGiftCard(giftCard, ...)

disableGiftCard(giftCardId, ...)

searchGiftCards(query, ...)
}
\arguments{
\item{...}{additional request parameters}

\item{giftCardId}{a GiftCard id number}

\item{giftCard}{a list containing GiftCard fields}

\item{query}{a query string to use when searching GiftCards}
}
\value{
a list containing a GiftCard fields or a list of GiftCards
}
\description{
GiftCard Functions
}
\references{
Shopify API GiftCard documentation: \url{https://help.shopify.com/api/reference/products/product}
}
