% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GraphQL.R
\name{GraphQL}
\alias{GraphQL}
\alias{graphQlQuery}
\title{Query the GraphQL Admin API}
\usage{
graphQlQuery(query, ..., includeExtensions = FALSE)
}
\arguments{
\item{query}{a literal GraphQL query string}

\item{...}{additional request parameters}

\item{includeExtensions}{should the extensions portion of the graphql query results be included in the return object. Default is \code{FALSE}}
}
\value{
The query results
}
\description{
Query the GraphQL Admin API
}
