% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InventoryItem.R
\name{InventoryItem}
\alias{InventoryItem}
\alias{getInventoryItems}
\alias{getInventoryItem}
\alias{modifyInventoryItem}
\title{InventoryItem Functions}
\usage{
getInventoryItems(inventoryItemIds, ...)

getInventoryItem(inventoryItemId, ...)

modifyInventoryItem(inventoryItem, ...)
}
\arguments{
\item{inventoryItemIds}{a vector of no more than 100 inventory item ids}

\item{...}{additional request parameters}

\item{inventoryItemId}{a InventoryItem id number}

\item{inventoryItem}{a list containing InventoryItem fields}
}
\value{
a list containing a InventoryItem fields or a list of InventoryItems
}
\description{
InventoryItem Functions
}
\references{
Shopify API InventoryItem documentation: \url{https://help.shopify.com/api/reference/inventory/inventoryitem}
}
