% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Location.R
\name{Location}
\alias{Location}
\alias{getLocations}
\alias{getLocation}
\alias{getLocationsCount}
\alias{getLocationInventoryLevels}
\title{Location Functions}
\usage{
getLocations(...)

getLocation(locationId, ...)

getLocationsCount(...)

getLocationInventoryLevels(locationId, ...)
}
\arguments{
\item{...}{additional request parameters}

\item{locationId}{a Location id}
}
\value{
a list containing a Location fields or a list of Locations
}
\description{
Location Functions
}
\references{
Shopify API Location documentation: \url{https://help.shopify.com/api/reference/inventory/location}
}
