% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MarketingEvent.R
\name{MarketingEvent}
\alias{MarketingEvent}
\alias{getMarketingEvents}
\alias{getMarketingEventsCount}
\alias{getMarketingEvent}
\alias{createMarketingEvent}
\alias{updateMarketingEvent}
\alias{deleteMarketingEvent}
\alias{createMarketingEventEngagements}
\title{MarketingEvent Functions}
\usage{
getMarketingEvents(...)

getMarketingEventsCount(...)

getMarketingEvent(marketingEventId, ...)

createMarketingEvent(marketingEvent, ...)

updateMarketingEvent(marketingEvent, ...)

deleteMarketingEvent(marketingEventId, ...)

createMarketingEventEngagements(marketingEventId, engagements, ...)
}
\arguments{
\item{...}{additional request parameters}

\item{marketingEventId}{a MarketingEvent id number}

\item{marketingEvent}{a list containing MarketingEvent fields}

\item{engagements}{a list of one or more engagements}
}
\value{
a list containing a MarketingEvent fields or a list of MarketingEvents
}
\description{
MarketingEvent Functions
}
\references{
Shopify API MarketingEvent documentation: \url{https://help.shopify.com/api/reference/marketingevent}
}
