% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MetaField.R
\name{Metafield}
\alias{Metafield}
\alias{getMetafields}
\alias{getMetafieldsCount}
\alias{getMetafield}
\alias{createMetafield}
\alias{modifyMetafield}
\alias{deleteMetafield}
\title{Metafield Functions}
\usage{
getMetafields(resourceName, resourceId = NULL, ...)

getMetafieldsCount(resourceName, resourceId = NULL, ...)

getMetafield(metafieldId, ...)

createMetafield(resourceName, resourceId = NULL, metafield, ...)

modifyMetafield(metafield, ...)

deleteMetafield(metafieldId, ...)
}
\arguments{
\item{resourceName}{the name of a resource e.g. \code{"products", "smart_collections", "product_image"}}

\item{resourceId}{the id number of the resource, if applicable (for example Shop has no id)}

\item{...}{additional request parameters}

\item{metafieldId}{a Metafield id number}

\item{metafield}{a list containing Metafield fields}
}
\value{
a list containing a Metafield fields or a list of Metafields
}
\description{
Metafield Functions
}
\references{
Shopify API Metafield documentation: \url{https://help.shopify.com/api/reference/metafield}
}
