% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Order.R
\name{Order}
\alias{Order}
\alias{getOrders}
\alias{getOrder}
\alias{getOrdersCount}
\alias{closeOrder}
\alias{openOrder}
\alias{cancelOrder}
\alias{createOrder}
\alias{modifyOrder}
\alias{deleteOrder}
\title{Order Functions}
\usage{
getOrders(...)

getOrder(orderId, ...)

getOrdersCount(...)

closeOrder(orderId, ...)

openOrder(orderId, ...)

cancelOrder(orderId, ...)

createOrder(order, ...)

modifyOrder(order, ...)

deleteOrder(orderId, ...)
}
\arguments{
\item{...}{additional request parameters}

\item{orderId}{a Order id number}

\item{order}{a list containing Order fields}
}
\value{
a list containing a Order fields or a list of Orders
}
\description{
Order Functions
}
\references{
Shopify API Order documentation: \url{https://help.shopify.com/api/reference/orders/order}
}
