% Generated by roxygen2 (4.0.1): do not edit by hand
\name{OrderRisks}
\alias{OrderRisks}
\alias{createOrderRisk}
\alias{deleteOrderRisk}
\alias{getOrderRisk}
\alias{getOrderRisks}
\alias{modifyOrderRisk}
\title{OrderRisks Functions}
\usage{
createOrderRisk(orderId, orderRisk, ...)

getOrderRisks(orderId, ...)

getOrderRisk(orderId, orderRiskId, ...)

modifyOrderRisk(orderId, orderRisk, ...)

deleteOrderRisk(orderId, orderRiskId, ...)
}
\arguments{
\item{orderId}{an Order id number}

\item{orderRiskId}{a OrderRisks id number}

\item{orderRisk}{a list containing OrderRisks fields}

\item{...}{additional request parameters}
}
\value{
a list containing a OrderRisks fields or a list of OrderRiskss
}
\description{
OrderRisks Functions
}
\references{
Shopify API OrderRisks documentation: \url{http://docs.shopify.com/api/order_risks}
}

