% Generated by roxygen2 (4.0.1): do not edit by hand
\name{Page}
\alias{Page}
\alias{createPage}
\alias{deletePage}
\alias{getPage}
\alias{getPages}
\alias{getPagesCount}
\alias{modifyPage}
\title{Page Functions}
\usage{
getPages(...)

getPagesCount(...)

getPage(pageId, ...)

createPage(page, ...)

modifyPage(page, ...)

deletePage(pageId, ...)
}
\arguments{
\item{pageId}{a Page id number}

\item{page}{a list containing Page fields}

\item{...}{additional request parameters}
}
\value{
a list containing a Page fields or a list of Pages
}
\description{
Page Functions
}
\references{
Shopify API Page documentation: \url{http://docs.shopify.com/api/page}
}

