% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Policy.R
\name{Policy}
\alias{Policy}
\alias{getPolicies}
\title{Policy Functions}
\usage{
getPolicies(...)
}
\arguments{
\item{...}{additional request parameters}
}
\value{
a list of Policies
}
\description{
Policy Functions
}
\references{
Shopify API Policy documentation: \url{https://help.shopify.com/api/reference/store-properties/policy}
}
