% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PriceRule.R
\name{PriceRule}
\alias{PriceRule}
\alias{getPriceRules}
\alias{getPriceRulesCount}
\alias{getPriceRule}
\alias{createPriceRule}
\alias{modifyPriceRule}
\alias{deletePriceRule}
\title{PriceRule Functions}
\usage{
getPriceRules(...)

getPriceRulesCount(...)

getPriceRule(priceRuleId, ...)

createPriceRule(priceRule, ...)

modifyPriceRule(priceRule, ...)

deletePriceRule(priceRuleId, ...)
}
\arguments{
\item{...}{additional request parameters}

\item{priceRuleId}{a PriceRule id number}

\item{priceRule}{a list containing PriceRule fields}
}
\value{
a list containing a PriceRule fields or a list of PriceRules
}
\description{
PriceRule Functions
}
\references{
Shopify API PriceRule documentation: \url{https://help.shopify.com/api/reference/discounts/pricerule}
}
