% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Product.R
\name{Product}
\alias{Product}
\alias{getProducts}
\alias{getProductsCount}
\alias{getProduct}
\alias{createProduct}
\alias{modifyProduct}
\alias{deleteProduct}
\title{Product Functions}
\usage{
getProducts(...)

getProductsCount(...)

getProduct(productId, ...)

createProduct(product, ...)

modifyProduct(product, ...)

deleteProduct(productId, ...)
}
\arguments{
\item{...}{additional request parameters}

\item{productId}{a Product id number}

\item{product}{a list containing Product fields}
}
\value{
a list containing a Product fields or a list of Products
}
\description{
Product Functions
}
\references{
Shopify API Product documentation: \url{https://help.shopify.com/api/reference/products/product}
}
