% Generated by roxygen2 (4.0.1): do not edit by hand
\name{ProductImage}
\alias{ProductImage}
\alias{createProductImage}
\alias{deleteProductImage}
\alias{getProductImage}
\alias{getProductImages}
\alias{getProductImagesCount}
\alias{modifyProductImage}
\title{ProductImage Functions}
\usage{
getProductImages(productId, ...)

getProductImagesCount(productId, ...)

getProductImage(productId, imageId, ...)

createProductImage(productId, image, ..., filepath)

modifyProductImage(productId, image, ..., filepath)

deleteProductImage(productId, imageId, ...)
}
\arguments{
\item{productId}{a Product id number}

\item{filepath}{the filepath to an image to upload}

\item{imageId}{a ProductImage id number}

\item{image}{a list containing ProductImage fields}

\item{...}{additional request parameters}
}
\value{
a list containing a ProductImage fields or a list of ProductImages
}
\description{
ProductImage Functions
}
\references{
Shopify API ProductImage documentation: \url{http://docs.shopify.com/api/product_image}
}

