% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProductListing.R
\name{ProductListing}
\alias{ProductListing}
\alias{getProductListings}
\alias{getProductListingProducts}
\alias{getProductListingsCount}
\alias{getProductListing}
\alias{createProductListing}
\alias{deleteProductListing}
\title{ProductListing Functions}
\usage{
getProductListings(...)

getProductListingProducts(...)

getProductListingsCount(...)

getProductListing(productListingId, ...)

createProductListing(productId, ...)

deleteProductListing(productListingId, ...)
}
\arguments{
\item{...}{additional request parameters}

\item{productListingId}{a ProductListing id}

\item{productId}{a \code{\link{Product}} id}
}
\value{
a list containing a ProductListing fields or a list of ProductListings
}
\description{
ProductListing Functions
}
\references{
Shopify API ProductListing documentation: \url{https://help.shopify.com/api/reference/sales-channels/productlisting}
}
