% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProductVariant.R
\name{ProductVariant}
\alias{ProductVariant}
\alias{getProductVariants}
\alias{getProductVariantsCount}
\alias{getProductVariant}
\alias{createProductVariant}
\alias{modifyProductVariant}
\alias{deleteProductVariant}
\title{ProductVariant Functions}
\usage{
getProductVariants(productId, ...)

getProductVariantsCount(productId, ...)

getProductVariant(variantId, ...)

createProductVariant(productId, variant, ...)

modifyProductVariant(productId, variant, ...)

deleteProductVariant(productId, variantId, ...)
}
\arguments{
\item{productId}{a \code{\link{Product}} id number}

\item{...}{additional request parameters}

\item{variantId}{a ProductVariant id number}

\item{variant}{a list containing ProductVariant fields}
}
\value{
a list containing a ProductVariant fields or a list of ProductVariants
}
\description{
ProductVariant Functions
}
\references{
Shopify API ProductVariant documentation: \url{https://help.shopify.com/api/reference/products/product_variant}
}
