% Generated by roxygen2 (4.0.1): do not edit by hand
\name{Province}
\alias{Province}
\alias{getProvince}
\alias{getProvinces}
\alias{getProvincesCount}
\alias{modifyProvince}
\title{Province Functions}
\usage{
getProvinces(countryId, ...)

getProvincesCount(countryId, ...)

getProvince(countryId, provinceId, ...)

modifyProvince(countryId, province, ...)
}
\arguments{
\item{countryId}{a Country id number}

\item{provinceId}{a Province id number}

\item{province}{a list containing Province fields}

\item{...}{additional request parameters}
}
\value{
a list containing a Province fields or a list of Provinces
}
\description{
Province Functions
}
\references{
Shopify API Province documentation: \url{http://docs.shopify.com/api/province}
}

