% Generated by roxygen2 (4.0.1): do not edit by hand
\name{RecurringApplicationCharge}
\alias{RecurringApplicationCharge}
\alias{activateRecurringCharge}
\alias{cancelRecurringCharge}
\alias{createRecurringCharge}
\alias{getRecurringCharge}
\alias{getRecurringCharges}
\title{RecurringApplicationCharge Functions}
\usage{
createRecurringCharge(charge, ...)

getRecurringCharge(chargeId, ...)

getRecurringCharges(...)

activateRecurringCharge(charge, ...)

cancelRecurringCharge(chargeId, ...)
}
\arguments{
\item{chargeId}{a RecurringApplicationCharge id number}

\item{charge}{a list containing RecurringApplicationCharge fields}

\item{...}{additional request parameters}
}
\value{
a list containing a RecurringApplicationCharge fields or a list of RecurringApplicationCharges
}
\description{
RecurringApplicationCharge Functions
}
\references{
Shopify API RecurringApplicationCharge documentation: \url{http://docs.shopify.com/api/recurringapplicationcharge}
}

