% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Redirect.R
\name{Redirect}
\alias{Redirect}
\alias{getRedirects}
\alias{getRedirectsCount}
\alias{getRedirect}
\alias{createRedirect}
\alias{modifyRedirect}
\alias{deleteRedirect}
\title{Redirect Functions}
\usage{
getRedirects(...)

getRedirectsCount(...)

getRedirect(redirectId, ...)

createRedirect(redirect, ...)

modifyRedirect(redirect, ...)

deleteRedirect(redirectId, ...)
}
\arguments{
\item{...}{additional request parameters}

\item{redirectId}{a Redirect id number}

\item{redirect}{a list containing Redirect fields}
}
\value{
a list containing a Redirect fields or a list of Redirects
}
\description{
Redirect Functions
}
\references{
Shopify API Redirect documentation: \url{https://help.shopify.com/api/reference/online_store/redirect}
}
