% Generated by roxygen2 (4.0.1): do not edit by hand
\name{Refund}
\alias{Refund}
\alias{getRefund}
\title{Refund Functions}
\usage{
getRefund(orderId, refundId, ...)
}
\arguments{
\item{orderId}{an Order id number}

\item{refundId}{a Refund id number}

\item{...}{additional request parameters}
}
\value{
a list corresponding to a Refund
}
\description{
Refund Functions
}
\references{
Shopify API Refund documentation: \url{http://docs.shopify.com/api/refund}
}

