% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Report.R
\name{Report}
\alias{Report}
\alias{getReports}
\alias{getReport}
\alias{createReport}
\alias{modifyReport}
\alias{deleteReport}
\title{Report Functions}
\usage{
getReports(...)

getReport(reportId, ...)

createReport(report, ...)

modifyReport(report, ...)

deleteReport(reportId, ...)
}
\arguments{
\item{...}{additional request parameters}

\item{reportId}{a Report id number}

\item{report}{a list containing Report fields}
}
\value{
a list containing a Report fields or a list of Reports
}
\description{
Report Functions
}
\references{
Shopify API Report documentation: \url{https://help.shopify.com/api/reference/analytics/report}
}
