% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ResourceFeedback.R
\name{ResourceFeedback}
\alias{ResourceFeedback}
\alias{createResourceFeedback}
\alias{getResourceFeedbacks}
\title{ResourceFeedback Functions}
\usage{
createResourceFeedback(resourceFeedback, ...)

getResourceFeedbacks(...)
}
\arguments{
\item{resourceFeedback}{a list containing ResourceFeedback fields}

\item{...}{additional request parameters}
}
\value{
a list containing a ResourceFeedback fields or a list of ResourceFeedbacks
}
\description{
ResourceFeedback Functions
}
\references{
Shopify API ResourceFeedback documentation: \url{https://help.shopify.com/api/reference/sales-channels/resourcefeedback}
}
