% Generated by roxygen2 (4.0.1): do not edit by hand
\name{ScriptTag}
\alias{ScriptTag}
\alias{createScriptTag}
\alias{deleteScriptTag}
\alias{getScriptTag}
\alias{getScriptTags}
\alias{getScriptTagsCount}
\alias{modifyScriptTag}
\title{ScriptTag Functions}
\usage{
getScriptTags(...)

getScriptTagsCount(...)

getScriptTag(scriptTagId, ...)

createScriptTag(scriptTag, ...)

modifyScriptTag(scriptTag, ...)

deleteScriptTag(scriptTagId, ...)
}
\arguments{
\item{scriptTagId}{a ScriptTag id number}

\item{scriptTag}{a list containing ScriptTag fields}

\item{...}{additional request parameters}
}
\value{
a list containing a ScriptTag fields or a list of ScriptTags
}
\description{
ScriptTag Functions
}
\references{
Shopify API ScriptTag documentation: \url{http://docs.shopify.com/api/scripttag}
}

