% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ShippingZone.R
\name{ShippingZone}
\alias{ShippingZone}
\alias{getShippingZones}
\title{ShippingZone Functions}
\usage{
getShippingZones(...)
}
\arguments{
\item{...}{additional request parameters}
}
\value{
a list of ShippingZones
}
\description{
ShippingZone Functions
}
\references{
Shopify API ShippingZone documentation: \url{https://help.shopify.com/api/reference/store-properties/shippingzone}
}
