% Generated by roxygen2 (4.0.1): do not edit by hand
\name{Shop}
\alias{Shop}
\alias{getShop}
\title{Shop Functions}
\usage{
getShop(...)
}
\arguments{
\item{...}{additional request parameters}
}
\value{
a list corresponding to a Shop
}
\description{
Shop Functions
}
\references{
Shopify API Shop documentation: \url{http://docs.shopify.com/api/shop}
}

