% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{package}
\name{shopifyr}
\alias{shopifyr}
\alias{shopifyr-package}
\title{R Interface to the Shopify API}
\description{
This package aims to provide an easy-to-use R interface to the Shopify API. Before using this
package, it would be wise to become familiar with the official Shopify API documentation. You
can find it at \url{http://docs.shopify.com/api}.
}
\details{
The workhorse of the package is the \code{\link{ShopifyShop}} class.
}
\author{
Charlie Friedemann
}

