% Generated by roxygen2 (4.0.1): do not edit by hand
\name{SmartCollection}
\alias{SmartCollection}
\alias{createSmartCollection}
\alias{deleteSmartCollection}
\alias{getSmartCollection}
\alias{getSmartCollections}
\alias{getSmartCollectionsCount}
\alias{modifySmartCollection}
\alias{orderSmartCollection}
\title{SmartCollection Functions}
\usage{
getSmartCollections(...)

getSmartCollectionsCount(...)

getSmartCollection(smartCollectionId, ...)

createSmartCollection(smartCollection, ...)

modifySmartCollection(smartCollection, ...)

orderSmartCollection(smartCollectionId, productIds, ...)

deleteSmartCollection(smartCollectionId, ...)
}
\arguments{
\item{productIds}{a vector of Product ids in the desired sort order}

\item{smartCollectionId}{a SmartCollection id number}

\item{smartCollection}{a list containing SmartCollection fields}

\item{...}{additional request parameters}
}
\value{
a list containing a SmartCollection fields or a list of SmartCollections
}
\description{
SmartCollection Functions
}
\references{
Shopify API SmartCollection documentation: \url{http://docs.shopify.com/api/smartcollection}
}

