% Generated by roxygen2 (4.0.1): do not edit by hand
\name{Theme}
\alias{Theme}
\alias{createTheme}
\alias{deleteTheme}
\alias{getTheme}
\alias{getThemes}
\alias{modifyTheme}
\title{Theme Functions}
\usage{
getThemes(...)

getTheme(themeId, ...)

createTheme(theme, ...)

modifyTheme(theme, ...)

deleteTheme(themeId, ...)
}
\arguments{
\item{themeId}{a Theme id number}

\item{theme}{a list containing Theme fields}

\item{...}{additional request parameters}
}
\value{
a list containing a Theme fields or a list of Themes
}
\description{
Theme Functions
}
\references{
Shopify API Theme documentation: \url{http://docs.shopify.com/api/theme}
}

