% Generated by roxygen2 (4.0.1): do not edit by hand
\name{Transaction}
\alias{Transaction}
\alias{createTransaction}
\alias{getTransaction}
\alias{getTransactions}
\alias{getTransactionsCount}
\title{Transaction Functions}
\usage{
getTransactions(orderId, ...)

getTransactionsCount(orderId, ...)

getTransaction(orderId, transactionId, ...)

createTransaction(orderId, transaction, ...)
}
\arguments{
\item{orderId}{an Order id number}

\item{transactionId}{a Transaction id number}

\item{transaction}{a list containing Transaction fields}

\item{...}{additional request parameters}
}
\value{
a list containing a Transaction fields or a list of Transactions
}
\description{
Transaction Functions
}
\references{
Shopify API Transaction documentation: \url{http://docs.shopify.com/api/transaction}
}

