% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UsageCharge.R
\name{UsageCharge}
\alias{UsageCharge}
\alias{createUsageCharge}
\alias{getUsageCharge}
\alias{getUsageCharges}
\title{UsageCharge Functions}
\usage{
createUsageCharge(recurringApplicationChargeId, usageCharge, ...)

getUsageCharge(recurringApplicationChargeId, usageChargeId, ...)

getUsageCharges(recurringApplicationChargeId, ...)
}
\arguments{
\item{recurringApplicationChargeId}{a \code{\link{RecurringApplicationCharge}} id}

\item{usageCharge}{a list containing UsageCharge fields}

\item{...}{additional request parameters}

\item{usageChargeId}{a UsageCharge id number}
}
\value{
a list containing a UsageCharge fields or a list of UsageCharges
}
\description{
UsageCharge Functions
}
\references{
Shopify API UsageCharge documentation: \url{https://help.shopify.com/api/reference/billing/usagecharge}
}
