% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/User.R
\name{User}
\alias{User}
\alias{getUsers}
\alias{getUser}
\alias{getCurrentUser}
\title{User Functions}
\usage{
getUsers(...)

getUser(userId, ...)

getCurrentUser(...)
}
\arguments{
\item{...}{additional request parameters}

\item{userId}{a User id}
}
\value{
a list containing a User fields or a list of Users
}
\description{
User Functions
}
\note{
This functionality is only available to Shopify Plus stores
}
\references{
Shopify API User documentation: \url{https://help.shopify.com/api/reference/plus/user}
}
