% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Webhook.R
\name{Webhook}
\alias{Webhook}
\alias{getWebhooks}
\alias{getWebhooksCount}
\alias{getWebhook}
\alias{createWebhook}
\alias{modifyWebhook}
\alias{deleteWebhook}
\title{Webhook Functions}
\usage{
getWebhooks(...)

getWebhooksCount(...)

getWebhook(webhookId, ...)

createWebhook(webhook, ...)

modifyWebhook(webhook, ...)

deleteWebhook(webhookId, ...)
}
\arguments{
\item{...}{additional request parameters}

\item{webhookId}{a Webhook id number}

\item{webhook}{a list containing Webhook fields}
}
\value{
a list containing a Webhook fields or a list of Webhooks
}
\description{
Webhook Functions
}
\references{
Shopify API Webhook documentation: \url{https://help.shopify.com/api/reference/events/webhook}
}
