% Generated by roxygen2 (4.0.1): do not edit by hand
\name{ApplicationCharge}
\alias{ApplicationCharge}
\alias{activateApplicationCharge}
\alias{createApplicationCharge}
\alias{getApplicationCharge}
\alias{getApplicationCharges}
\title{ApplicationCharge Functions}
\usage{
createApplicationCharge(charge, ...)

getApplicationCharge(chargeId, ...)

getApplicationCharges(...)

activateApplicationCharge(charge, ...)
}
\arguments{
\item{chargeId}{a ApplicationCharge id number}

\item{charge}{a list containing ApplicationCharge fields}

\item{...}{additional request parameters}
}
\value{
a list containing a ApplicationCharge fields or a list of ApplicationCharges
}
\description{
ApplicationCharge Functions
}
\references{
Shopify API ApplicationCharge documentation: \url{http://docs.shopify.com/api/applicationcharge}
}

