% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Article.R
\name{Article}
\alias{Article}
\alias{getArticles}
\alias{getArticlesCount}
\alias{getArticle}
\alias{createArticle}
\alias{modifyArticle}
\alias{getArticleAuthors}
\alias{getArticleTags}
\alias{deleteArticle}
\title{Article Functions}
\usage{
getArticles(blogId, ...)

getArticlesCount(blogId, ...)

getArticle(blogId, articleId, ...)

createArticle(blogId, article, ...)

modifyArticle(blogId, article, ...)

getArticleAuthors(...)

getArticleTags(...)

deleteArticle(blogId, articleId, ...)
}
\arguments{
\item{blogId}{a \code{\link{Blog}} id number}

\item{...}{additional request parameters}

\item{articleId}{an Article id number}

\item{article}{a list containing Article fields}
}
\value{
a list containing an Article fields or a list of Articles
}
\description{
Article Functions
}
\references{
Shopify API Article documentation: \url{https://help.shopify.com/api/reference/online_store/blog_article}
}
