% Generated by roxygen2 (4.0.1): do not edit by hand
\name{Blog}
\alias{Blog}
\alias{createBlog}
\alias{deleteBlog}
\alias{getBlog}
\alias{getBlogs}
\alias{getBlogsCount}
\alias{modifyBlog}
\title{Blog Functions}
\usage{
getBlogs(...)

getBlogsCount(...)

getBlog(blogId, ...)

createBlog(blog, ...)

modifyBlog(blog, ...)

deleteBlog(blogId, ...)
}
\arguments{
\item{blogId}{a Blog id number}

\item{blog}{a list containing Blog fields}

\item{...}{additional request parameters}
}
\value{
a list containing a Blog fields or a list of Blogs
}
\description{
Blog Functions
}
\references{
Shopify API Blog documentation: \url{http://docs.shopify.com/api/blog}
}

