% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CarrierService.R
\name{CarrierService}
\alias{CarrierService}
\alias{createCarrierService}
\alias{modifyCarrierService}
\alias{getCarrierServices}
\alias{getCarrierService}
\alias{deleteCarrierService}
\title{CarrierService Functions}
\usage{
createCarrierService(service, ...)

modifyCarrierService(service, ...)

getCarrierServices(...)

getCarrierService(serviceId, ...)

deleteCarrierService(serviceId, ...)
}
\arguments{
\item{service}{a list containing CarrierService fields}

\item{...}{additional request parameters}

\item{serviceId}{a CarrierService id number}
}
\value{
a list containing a CarrierService fields or a list of CarrierServices
}
\description{
CarrierService Functions
}
\references{
Shopify API CarrierService documentation: \url{https://help.shopify.com/api/reference/shipping_and_fulfillment/carrierservice}
}
