% Generated by roxygen2 (4.0.1): do not edit by hand
\name{Checkout}
\alias{Checkout}
\alias{getCheckouts}
\alias{getCheckoutsCount}
\title{Checkout Functions}
\usage{
getCheckoutsCount(...)

getCheckouts(...)
}
\arguments{
\item{...}{additional request parameters}
}
\value{
a list of Checkouts
}
\description{
Checkout Functions
}
\references{
Shopify API Checkout documentation: \url{http://docs.shopify.com/api/checkout}
}

