% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Collect.R
\name{Collect}
\alias{Collect}
\alias{createCollect}
\alias{deleteCollect}
\alias{getCollects}
\alias{getCollectsCount}
\alias{getCollect}
\title{Collect Functions}
\usage{
createCollect(collect, ...)

deleteCollect(collectId, ...)

getCollects(...)

getCollectsCount(...)

getCollect(collectId, ...)
}
\arguments{
\item{collect}{a list containing Collect fields}

\item{...}{additional request parameters}

\item{collectId}{a Collect id number}
}
\value{
a list containing a Collect fields or a list of Collects
}
\description{
Collect Functions
}
\references{
Shopify API Collect documentation: \url{https://help.shopify.com/api/reference/products/collect}
}
