% Generated by roxygen2 (4.0.1): do not edit by hand
\name{Country}
\alias{Country}
\alias{createCountry}
\alias{deleteCountry}
\alias{getCountries}
\alias{getCountriesCount}
\alias{getCountry}
\alias{modifyCountry}
\title{Country Functions}
\usage{
getCountries(...)

getCountriesCount(...)

getCountry(countryId, ...)

createCountry(country, ...)

modifyCountry(country, ...)

deleteCountry(countryId, ...)
}
\arguments{
\item{countryId}{a Country id number}

\item{country}{a list containing Country fields}

\item{...}{additional request parameters}
}
\value{
a list containing a Country fields or a list of Countrys
}
\description{
Country Functions
}
\references{
Shopify API Country documentation: \url{http://docs.shopify.com/api/country}
}

