% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CustomCollection.R
\name{CustomCollection}
\alias{CustomCollection}
\alias{getCustomCollections}
\alias{getCustomCollectionsCount}
\alias{getCustomCollection}
\alias{createCustomCollection}
\alias{modifyCustomCollection}
\alias{deleteCustomCollection}
\title{CustomCollection Functions}
\usage{
getCustomCollections(...)

getCustomCollectionsCount(...)

getCustomCollection(customCollectionId, ...)

createCustomCollection(customCollection, ...)

modifyCustomCollection(customCollection, ...)

deleteCustomCollection(customCollectionId, ...)
}
\arguments{
\item{...}{additional request parameters}

\item{customCollectionId}{a CustomCollection id number}

\item{customCollection}{a list containing CustomCollection fields}
}
\value{
a list containing a CustomCollection fields or a list of CustomCollections
}
\description{
CustomCollection Functions
}
\references{
Shopify API CustomCollection documentation: \url{https://help.shopify.com/api/reference/products/customcollection}
}
