% Generated by roxygen2 (4.0.1): do not edit by hand
\name{Customer}
\alias{Customer}
\alias{createCustomer}
\alias{deleteCustomer}
\alias{getCustomer}
\alias{getCustomerOrders}
\alias{getCustomers}
\alias{getCustomersCount}
\alias{modifyCustomer}
\alias{searchCustomers}
\title{Customer Functions}
\usage{
getCustomers(...)

searchCustomers(query, ...)

getCustomer(customerId, ...)

createCustomer(customer, ...)

modifyCustomer(customer, ...)

deleteCustomer(customerId, ...)

getCustomersCount(...)

getCustomerOrders(customerId, ...)
}
\arguments{
\item{query}{a query string to use when searching Customer records}

\item{customerId}{a Customer id number}

\item{customer}{a list containing Customer fields}

\item{...}{additional request parameters}
}
\value{
a list containing a Customer fields or a list of Customers
}
\description{
Customer Functions
}
\references{
Shopify API Customer documentation: \url{http://docs.shopify.com/api/customer}
}

