% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Event.R
\name{Event}
\alias{Event}
\alias{getEvents}
\alias{getEvent}
\alias{getEventsCount}
\title{Event Functions}
\usage{
getEvents(...)

getEvent(eventId, ...)

getEventsCount(...)
}
\arguments{
\item{...}{additional request parameters}

\item{eventId}{a Event id number}
}
\value{
a list containing a Event fields or a list of Events
}
\description{
Event Functions
}
\references{
Shopify API Event documentation: \url{https://help.shopify.com/api/reference/events/event}
}
