% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Fulfillment.R
\name{Fulfillment}
\alias{Fulfillment}
\alias{getFulfillments}
\alias{getFulfillmentsCount}
\alias{getFulfillment}
\alias{createFulfillment}
\alias{modifyFulfillment}
\alias{completeFulfillment}
\alias{cancelFulfillment}
\title{Fulfillment Functions}
\usage{
getFulfillments(orderId, ...)

getFulfillmentsCount(orderId, ...)

getFulfillment(orderId, fulfillmentId, ...)

createFulfillment(orderId, fulfillment, ...)

modifyFulfillment(orderId, fulfillment, ...)

completeFulfillment(orderId, fulfillmentId, ...)

cancelFulfillment(orderId, fulfillmentId, ...)
}
\arguments{
\item{orderId}{an \code{\link{Order}}id number}

\item{...}{additional request parameters}

\item{fulfillmentId}{a Fulfillment id number}

\item{fulfillment}{a list containing Fulfillment fields}
}
\value{
a list containing a Fulfillment fields or a list of Fulfillments
}
\description{
Fulfillment Functions
}
\references{
Shopify API Fulfillment documentation: \url{https://help.shopify.com/api/reference/shipping_and_fulfillment/fulfillment}
}
