% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FulfillmentService.R
\name{FulfillmentService}
\alias{FulfillmentService}
\alias{getFulfillmentServices}
\alias{createFulfillmentService}
\alias{getFulfillmentService}
\alias{modifyFulfillmentService}
\alias{deleteFulfillmentService}
\title{FulfillmentService Functions}
\usage{
getFulfillmentServices(...)

createFulfillmentService(fulfillmentService, ...)

getFulfillmentService(fulfillmentServiceId, ...)

modifyFulfillmentService(fulfillmentService, ...)

deleteFulfillmentService(fulfillmentServiceId, ...)
}
\arguments{
\item{...}{additional request parameters}

\item{fulfillmentService}{a list containing FulfillmentService fields}

\item{fulfillmentServiceId}{a FulfillmentService id number}
}
\value{
a list containing a FulfillmentService fields or a list of FulfillmentServices
}
\description{
FulfillmentService Functions
}
\references{
Shopify API FulfillmentService documentation: \url{https://help.shopify.com/api/reference/shipping_and_fulfillment/fulfillmentservice}
}
