% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InventoryLevel.R
\name{InventoryLevel}
\alias{InventoryLevel}
\alias{getInventoryLevels}
\alias{modifyInventoryLevel}
\alias{deleteInventoryLevel}
\alias{connectInventoryItem}
\alias{setInventoryLevel}
\title{InventoryLevel Functions}
\usage{
getInventoryLevels(locationIds, inventoryItemIds, ...)

modifyInventoryLevel(inventoryLevel, ...)

deleteInventoryLevel(locationId, inventoryItemId, ...)

connectInventoryItem(locationId, inventoryItemId, ...)

setInventoryLevel(inventoryLevel, ...)
}
\arguments{
\item{locationIds}{a vector of Location ids}

\item{inventoryItemIds}{a vector of InventoryItem ids}

\item{...}{additional request parameters}

\item{inventoryLevel}{a list containing InventoryLevel fields}

\item{locationId}{a single Location id}

\item{inventoryItemId}{a single InventoryItem id}
}
\value{
a list containing a InventoryLevel fields or a list of InventoryLevels
}
\description{
InventoryLevel Functions
}
\references{
Shopify API InventoryLevel documentation: \url{https://help.shopify.com/api/reference/inventory/inventorylevel}
}
