% Generated by roxygen2 (4.0.1): do not edit by hand
\name{Location}
\alias{Location}
\alias{getLocation}
\alias{getLocations}
\title{Location Functions}
\usage{
getLocations(...)

getLocation(locationId, ...)
}
\arguments{
\item{locationId}{a Location id number}

\item{...}{additional request parameters}
}
\value{
a list containing a Location fields or a list of Locations
}
\description{
Location Functions
}
\references{
Shopify API Location documentation: \url{http://docs.shopify.com/api/location}
}

