% Generated by roxygen2 (4.0.1): do not edit by hand
\name{Order}
\alias{Order}
\alias{cancelOrder}
\alias{closeOrder}
\alias{createOrder}
\alias{deleteOrder}
\alias{getOrder}
\alias{getOrders}
\alias{getOrdersCount}
\alias{modifyOrder}
\alias{openOrder}
\title{Order Functions}
\usage{
getOrders(...)

getOrder(orderId, ...)

getOrdersCount(...)

closeOrder(orderId, ...)

openOrder(orderId, ...)

cancelOrder(orderId, ...)

createOrder(order, ...)

modifyOrder(order, ...)

deleteOrder(orderId, ...)
}
\arguments{
\item{orderId}{a Order id number}

\item{order}{a list containing Order fields}

\item{...}{additional request parameters}
}
\value{
a list containing a Order fields or a list of Orders
}
\description{
Order Functions
}
\references{
Shopify API Order documentation: \url{http://docs.shopify.com/api/order}
}

