% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OrderRisks.R
\name{OrderRisks}
\alias{OrderRisks}
\alias{createOrderRisk}
\alias{getOrderRisks}
\alias{getOrderRisk}
\alias{modifyOrderRisk}
\alias{deleteOrderRisk}
\title{OrderRisks Functions}
\usage{
createOrderRisk(orderId, orderRisk, ...)

getOrderRisks(orderId, ...)

getOrderRisk(orderId, orderRiskId, ...)

modifyOrderRisk(orderId, orderRisk, ...)

deleteOrderRisk(orderId, orderRiskId, ...)
}
\arguments{
\item{orderId}{an \code{\link{Order}} id number}

\item{orderRisk}{a list containing OrderRisks fields}

\item{...}{additional request parameters}

\item{orderRiskId}{a OrderRisks id number}
}
\value{
a list containing a OrderRisks fields or a list of OrderRiskss
}
\description{
OrderRisks Functions
}
\references{
Shopify API OrderRisks documentation: \url{https://help.shopify.com/api/reference/orders/order_risk}
}
