% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Page.R
\name{Page}
\alias{Page}
\alias{getPages}
\alias{getPagesCount}
\alias{getPage}
\alias{createPage}
\alias{modifyPage}
\alias{deletePage}
\title{Page Functions}
\usage{
getPages(...)

getPagesCount(...)

getPage(pageId, ...)

createPage(page, ...)

modifyPage(page, ...)

deletePage(pageId, ...)
}
\arguments{
\item{...}{additional request parameters}

\item{pageId}{a Page id number}

\item{page}{a list containing Page fields}
}
\value{
a list containing a Page fields or a list of Pages
}
\description{
Page Functions
}
\references{
Shopify API Page documentation: \url{https://help.shopify.com/api/reference/online_store/page}
}
