% Generated by roxygen2 (4.0.1): do not edit by hand
\name{Product}
\alias{Product}
\alias{createProduct}
\alias{deleteProduct}
\alias{getProduct}
\alias{getProducts}
\alias{getProductsCount}
\alias{modifyProduct}
\title{Product Functions}
\usage{
getProducts(...)

getProductsCount(...)

getProduct(productId, ...)

createProduct(product, ...)

modifyProduct(product, ...)

deleteProduct(productId, ...)
}
\arguments{
\item{productId}{a Product id number}

\item{product}{a list containing Product fields}

\item{...}{additional request parameters}
}
\value{
a list containing a Product fields or a list of Products
}
\description{
Product Functions
}
\references{
Shopify API Product documentation: \url{http://docs.shopify.com/api/product}
}

