% Generated by roxygen2 (4.0.1): do not edit by hand
\name{ProductVariant}
\alias{ProductVariant}
\alias{createProductVariant}
\alias{deleteProductVariant}
\alias{getProductVariant}
\alias{getProductVariants}
\alias{getProductVariantsCount}
\alias{modifyProductVariant}
\title{ProductVariant Functions}
\usage{
getProductVariants(productId, ...)

getProductVariantsCount(productId, ...)

getProductVariant(variantId, ...)

createProductVariant(productId, variant, ...)

modifyProductVariant(productId, variant, ...)

deleteProductVariant(productId, variantId, ...)
}
\arguments{
\item{productId}{a Product id number}

\item{variantId}{a ProductVariant id number}

\item{variant}{a list containing ProductVariant fields}

\item{...}{additional request parameters}
}
\value{
a list containing a ProductVariant fields or a list of ProductVariants
}
\description{
ProductVariant Functions
}
\references{
Shopify API ProductVariant documentation: \url{http://docs.shopify.com/api/product_variant}
}

