% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Province.R
\name{Province}
\alias{Province}
\alias{getProvinces}
\alias{getProvincesCount}
\alias{getProvince}
\alias{modifyProvince}
\title{Province Functions}
\usage{
getProvinces(countryId, ...)

getProvincesCount(countryId, ...)

getProvince(countryId, provinceId, ...)

modifyProvince(countryId, province, ...)
}
\arguments{
\item{countryId}{a Country id number}

\item{...}{additional request parameters}

\item{provinceId}{a Province id number}

\item{province}{a list containing Province fields}
}
\value{
a list containing a Province fields or a list of Provinces
}
\description{
Province Functions
}
\references{
Shopify API Province documentation: \url{https://help.shopify.com/api/reference/store_properties/province}
}
