% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RecurringApplicationCharge.R
\name{RecurringApplicationCharge}
\alias{RecurringApplicationCharge}
\alias{createRecurringCharge}
\alias{getRecurringCharge}
\alias{getRecurringCharges}
\alias{activateRecurringCharge}
\alias{cancelRecurringCharge}
\title{RecurringApplicationCharge Functions}
\usage{
createRecurringCharge(charge, ...)

getRecurringCharge(chargeId, ...)

getRecurringCharges(...)

activateRecurringCharge(charge, ...)

cancelRecurringCharge(chargeId, ...)
}
\arguments{
\item{charge}{a list containing RecurringApplicationCharge fields}

\item{...}{additional request parameters}

\item{chargeId}{a RecurringApplicationCharge id number}
}
\value{
a list containing a RecurringApplicationCharge fields or a list of RecurringApplicationCharges
}
\description{
RecurringApplicationCharge Functions
}
\references{
Shopify API RecurringApplicationCharge documentation: \url{https://help.shopify.com/api/reference/billing/recurringapplicationcharge}
}
