% Generated by roxygen2 (4.0.1): do not edit by hand
\name{Redirect}
\alias{Redirect}
\alias{createRedirect}
\alias{deleteRedirect}
\alias{getRedirect}
\alias{getRedirects}
\alias{getRedirectsCount}
\alias{modifyRedirect}
\title{Redirect Functions}
\usage{
getRedirects(...)

getRedirectsCount(...)

getRedirect(redirectId, ...)

createRedirect(redirect, ...)

modifyRedirect(redirect, ...)

deleteRedirect(redirectId, ...)
}
\arguments{
\item{redirectId}{a Redirect id number}

\item{redirect}{a list containing Redirect fields}

\item{...}{additional request parameters}
}
\value{
a list containing a Redirect fields or a list of Redirects
}
\description{
Redirect Functions
}
\references{
Shopify API Redirect documentation: \url{http://docs.shopify.com/api/redirect}
}

