% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Refund.R
\name{Refund}
\alias{Refund}
\alias{getRefund}
\title{Refund Functions}
\usage{
getRefund(orderId, refundId, ...)
}
\arguments{
\item{orderId}{an \code{\link{Order}} id number}

\item{refundId}{a Refund id number}

\item{...}{additional request parameters}
}
\value{
a list corresponding to a Refund
}
\description{
Refund Functions
}
\references{
Shopify API Refund documentation: \url{https://help.shopify.com/api/reference/orders/refund}
}
