% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ScriptTag.R
\name{ScriptTag}
\alias{ScriptTag}
\alias{getScriptTags}
\alias{getScriptTagsCount}
\alias{getScriptTag}
\alias{createScriptTag}
\alias{modifyScriptTag}
\alias{deleteScriptTag}
\title{ScriptTag Functions}
\usage{
getScriptTags(...)

getScriptTagsCount(...)

getScriptTag(scriptTagId, ...)

createScriptTag(scriptTag, ...)

modifyScriptTag(scriptTag, ...)

deleteScriptTag(scriptTagId, ...)
}
\arguments{
\item{...}{additional request parameters}

\item{scriptTagId}{a ScriptTag id number}

\item{scriptTag}{a list containing ScriptTag fields}
}
\value{
a list containing a ScriptTag fields or a list of ScriptTags
}
\description{
ScriptTag Functions
}
\references{
Shopify API ScriptTag documentation: \url{https://help.shopify.com/api/reference/online_store/scripttag}
}
