% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Shop.R
\name{Shop}
\alias{Shop}
\alias{getShop}
\title{Shop Functions}
\usage{
getShop(...)
}
\arguments{
\item{...}{additional request parameters}
}
\value{
a list corresponding to a Shop
}
\description{
Shop Functions
}
\references{
Shopify API Shop documentation: \url{https://help.shopify.com/api/reference/store_properties/shop}
}
