% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{data}
\name{ShopifyShop}
\alias{ShopifyShop}
\title{ShopifyShop class}
\usage{
ShopifyShop$new(shopName, apiKey, password, sharedSecret, quiet = FALSE)
}
\arguments{
\item{shopName}{the name of your shop, as in shop-name.myshopify.com}

\item{apiKey}{the API key authorized in the Shopify admin}

\item{password}{the Shopify API password to be used with the API key}

\item{sharedSecret}{an authorized shared secret to be used with the API key}

\item{quiet}{suppress output of API announcements}
}
\value{
a ShopifyShop object
}
\description{
The \code{ShopifyShop} class fully encapsulates the Shopify API. It is an \code{\link{R6}} class, and
as such is initialized via the \sQuote{new} function (see the example section for details).
}
\details{
In order to access the Shopify API, users will need a set of authorized API access credentials.
Generally, users will create these so-called \sQuote{private app} credentials in their Shopify store's
admin section. More information about how to do this can be found
\href{http://docs.shopify.com/api/tutorials/creating-a-private-app}{here}. Once the credentials are
created, they remain valid and can be used in subsequent R sessions.
}
\section{Fields}{

\describe{
\item{\code{shopInfo}}{information about the shop as returned by \code{\link{getShop}}}
}}
\section{Public Functions}{

\describe{
\item{ApplicationCharge functions}{\itemize{
     \item \code{\link{createApplicationCharge}}
     \item \code{\link{getApplicationCharge}}
     \item \code{\link{getApplicationCharges}}
     \item \code{\link{activateApplicationCharge}}
}}
\item{Article functions}{\itemize{
     \item \code{\link{getArticles}}
     \item \code{\link{getArticlesCount}}
     \item \code{\link{getArticle}}
     \item \code{\link{createArticle}}
     \item \code{\link{modifyArticle}}
     \item \code{\link{getArticleAuthors}}
     \item \code{\link{getArticleTags}}
     \item \code{\link{deleteArticle}}
}}
\item{Asset functions}{\itemize{
     \item \code{\link{getAssets}}
     \item \code{\link{getAsset}}
     \item \code{\link{createAsset}}
     \item \code{\link{modifyAsset}}
     \item \code{\link{deleteAsset}}
}}
\item{Blog functions}{\itemize{
     \item \code{\link{getBlogs}}
     \item \code{\link{getBlogsCount}}
     \item \code{\link{getBlog}}
     \item \code{\link{createBlog}}
     \item \code{\link{modifyBlog}}
     \item \code{\link{deleteBlog}}
}}
\item{CarrierService functions}{\itemize{
     \item \code{\link{createCarrierService}}
     \item \code{\link{modifyCarrierService}}
     \item \code{\link{getCarrierServices}}
     \item \code{\link{getCarrierService}}
     \item \code{\link{deleteCarrierService}}
}}
\item{Checkout functions}{\itemize{
     \item \code{\link{getCheckoutsCount}}
     \item \code{\link{getCheckouts}}
}}
\item{Collect functions}{\itemize{
     \item \code{\link{createCollect}}
     \item \code{\link{deleteCollect}}
     \item \code{\link{getCollects}}
     \item \code{\link{getCollectsCount}}
     \item \code{\link{getCollect}}
}}
\item{Comment functions}{\itemize{
     \item \code{\link{getComments}}
     \item \code{\link{getCommentsCount}}
     \item \code{\link{getComment}}
     \item \code{\link{createComment}}
     \item \code{\link{modifyComment}}
     \item \code{\link{markCommentAsSpam}}
     \item \code{\link{markCommentAsNotSpam}}
     \item \code{\link{approveComment}}
     \item \code{\link{removeComment}}
     \item \code{\link{restoreComment}}
}}
\item{Country functions}{\itemize{
     \item \code{\link{getCountries}}
     \item \code{\link{getCountriesCount}}
     \item \code{\link{getCountry}}
     \item \code{\link{createCountry}}
     \item \code{\link{modifyCountry}}
     \item \code{\link{deleteCountry}}
}}
\item{CustomCollection functions}{\itemize{
     \item \code{\link{getCustomCollections}}
     \item \code{\link{getCustomCollectionsCount}}
     \item \code{\link{getCustomCollection}}
     \item \code{\link{createCustomCollection}}
     \item \code{\link{modifyCustomCollection}}
     \item \code{\link{deleteCustomCollection}}
}}
\item{Customer functions}{\itemize{
     \item \code{\link{getCustomers}}
     \item \code{\link{searchCustomers}}
     \item \code{\link{getCustomer}}
     \item \code{\link{createCustomer}}
     \item \code{\link{modifyCustomer}}
     \item \code{\link{deleteCustomer}}
     \item \code{\link{getCustomersCount}}
     \item \code{\link{getCustomerOrders}}
}}
\item{CustomerGroup & CustomerSavedSearch functions}{\itemize{
     \item \code{\link{getCustomerGroups}}
     \item \code{\link{getCustomerGroupsCount}}
     \item \code{\link{getCustomerGroup}}
     \item \code{\link{getCustomerGroupCustomers}}
     \item \code{\link{createCustomerGroup}}
     \item \code{\link{modifyCustomerGroup}}
     \item \code{\link{deleteCustomerGroup}}
}}
\item{Event functions}{\itemize{
     \item \code{\link{getEvents}}
     \item \code{\link{getEvent}}
     \item \code{\link{getEventsCount}}
}}
\item{Fulfillment functions}{\itemize{
     \item \code{\link{getFulfillments}}
     \item \code{\link{getFulfillmentsCount}}
     \item \code{\link{getFulfillment}}
     \item \code{\link{createFulfillment}}
     \item \code{\link{modifyFulfillment}}
     \item \code{\link{completeFulfillment}}
     \item \code{\link{cancelFulfillment}}
}}
\item{FulfillmentService functions}{\itemize{
     \item \code{\link{getFulfillmentServices}}
     \item \code{\link{createFulfillmentService}}
     \item \code{\link{getFulfillmentService}}
     \item \code{\link{modifyFulfillmentService}}
     \item \code{\link{deleteFulfillmentService}}
}}
\item{Location functions}{\itemize{
     \item \code{\link{getLocations}}
     \item \code{\link{getLocation}}
}}
\item{Metafield functions}{\itemize{
     \item \code{\link{getMetafields}}
     \item \code{\link{getMetafieldsCount}}
     \item \code{\link{getMetafield}}
     \item \code{\link{createMetafield}}
     \item \code{\link{modifyMetafield}}
     \item \code{\link{deleteMetafield}}
}}
\item{Order functions}{\itemize{
     \item \code{\link{getOrders}}
     \item \code{\link{getOrder}}
     \item \code{\link{getOrdersCount}}
     \item \code{\link{closeOrder}}
     \item \code{\link{openOrder}}
     \item \code{\link{cancelOrder}}
     \item \code{\link{createOrder}}
     \item \code{\link{modifyOrder}}
     \item \code{\link{deleteOrder}}
}}
\item{Order Risks functions}{\itemize{
     \item \code{\link{createOrderRisk}}
     \item \code{\link{getOrderRisks}}
     \item \code{\link{getOrderRisk}}
     \item \code{\link{modifyOrderRisk}}
     \item \code{\link{deleteOrderRisk}}
}}
\item{Page functions}{\itemize{
     \item \code{\link{getPages}}
     \item \code{\link{getPagesCount}}
     \item \code{\link{getPage}}
     \item \code{\link{createPage}}
     \item \code{\link{modifyPage}}
     \item \code{\link{deletePage}}
}}
\item{Product functions}{\itemize{
     \item \code{\link{getProducts}}
     \item \code{\link{getProductsCount}}
     \item \code{\link{getProduct}}
     \item \code{\link{createProduct}}
     \item \code{\link{modifyProduct}}
     \item \code{\link{deleteProduct}}
}}
\item{Product Image functions}{\itemize{
     \item \code{\link{getProductImages}}
     \item \code{\link{getProductImagesCount}}
     \item \code{\link{getProductImage}}
     \item \code{\link{createProductImage}}
     \item \code{\link{modifyProductImage}}
     \item \code{\link{deleteProductImage}}
}}
\item{Product Variant functions}{\itemize{
     \item \code{\link{getProductVariants}}
     \item \code{\link{getProductVariantsCount}}
     \item \code{\link{getProductVariant}}
     \item \code{\link{createProductVariant}}
     \item \code{\link{modifyProductVariant}}
     \item \code{\link{deleteProductVariant}}
}}
\item{Province functions}{\itemize{
     \item \code{\link{getProvinces}}
     \item \code{\link{getProvincesCount}}
     \item \code{\link{getProvince}}
     \item \code{\link{modifyProvince}}
}}
\item{RecurringApplicationCharge functions}{\itemize{
     \item \code{\link{createRecurringCharge}}
     \item \code{\link{getRecurringCharge}}
     \item \code{\link{getRecurringCharges}}
     \item \code{\link{activateRecurringCharge}}
     \item \code{\link{cancelRecurringCharge}}
}}
\item{Redirect functions}{\itemize{
     \item \code{\link{getRedirects}}
     \item \code{\link{getRedirectsCount}}
     \item \code{\link{getRedirect}}
     \item \code{\link{createRedirect}}
     \item \code{\link{modifyRedirect}}
     \item \code{\link{deleteRedirect}}
}}
\item{Refund functions}{\itemize{
     \item \code{\link{getRefund}}
}}
\item{ScriptTag functions}{\itemize{
     \item \code{\link{getScriptTags}}
     \item \code{\link{getScriptTagsCount}}
     \item \code{\link{getScriptTag}}
     \item \code{\link{createScriptTag}}
     \item \code{\link{modifyScriptTag}}
     \item \code{\link{deleteScriptTag}}
}}
\item{Shop functions}{\itemize{
     \item \code{\link{getShop}}
}}
\item{SmartCollection functions}{\itemize{
     \item \code{\link{getSmartCollections}}
     \item \code{\link{getSmartCollectionsCount}}
     \item \code{\link{getSmartCollection}}
     \item \code{\link{createSmartCollection}}
     \item \code{\link{modifySmartCollection}}
     \item \code{\link{orderSmartCollection}}
     \item \code{\link{deleteSmartCollection}}
}}
\item{Theme functions}{\itemize{
     \item \code{\link{getThemes}}
     \item \code{\link{getTheme}}
     \item \code{\link{createTheme}}
     \item \code{\link{modifyTheme}}
     \item \code{\link{deleteTheme}}
}}
\item{Transaction functions}{\itemize{
     \item \code{\link{getTransactions}}
     \item \code{\link{getTransactionsCount}}
     \item \code{\link{getTransaction}}
     \item \code{\link{createTransaction}}
}}
\item{Webhook functions}{\itemize{
     \item \code{\link{getWebhooks}}
     \item \code{\link{getWebhooksCount}}
     \item \code{\link{getWebhook}}
     \item \code{\link{createWebhook}}
     \item \code{\link{modifyWebhook}}
     \item \code{\link{deleteWebhook}}
}}
}
}
\examples{
\dontrun{
# Before you can access the API, you must create a ShopifyShop object using your credentials
shop <- ShopifyShop$new("yourstore.myshopify.com","api_key","password","shared_secret")

# Get all collections
collections <- shop$getCollections()

# Get customers named Bob from the United States
bobs <- shop$getCustomers("Bob country:United States")

# Create a new product
product <- list(title="The R Inferno",
                body="<b>The R Inferno</b><i>by Patrick Burns</i>",
                vendor="Burns Statistics",
                product_type="Paperback",
                images=list(list(src="http://www.burns-stat.com/wp-content/
                                      uploads/2012/12/R_inferno_cover.jpg")))
newProduct <- shop$createProduct(product)
}
}
\keyword{datasets}

