% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SmartCollection.R
\name{SmartCollection}
\alias{SmartCollection}
\alias{getSmartCollections}
\alias{getSmartCollectionsCount}
\alias{getSmartCollection}
\alias{createSmartCollection}
\alias{modifySmartCollection}
\alias{orderSmartCollection}
\alias{deleteSmartCollection}
\title{SmartCollection Functions}
\usage{
getSmartCollections(...)

getSmartCollectionsCount(...)

getSmartCollection(smartCollectionId, ...)

createSmartCollection(smartCollection, ...)

modifySmartCollection(smartCollection, ...)

orderSmartCollection(smartCollectionId, productIds, ...)

deleteSmartCollection(smartCollectionId, ...)
}
\arguments{
\item{...}{additional request parameters}

\item{smartCollectionId}{a SmartCollection id number}

\item{smartCollection}{a list containing SmartCollection fields}

\item{productIds}{a vector of Product ids in the desired sort order}
}
\value{
a list containing a SmartCollection fields or a list of SmartCollections
}
\description{
SmartCollection Functions
}
\references{
Shopify API SmartCollection documentation: \url{https://help.shopify.com/api/reference/products/smartcollection}
}
