% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Theme.R
\name{Theme}
\alias{Theme}
\alias{getThemes}
\alias{getTheme}
\alias{createTheme}
\alias{modifyTheme}
\alias{deleteTheme}
\title{Theme Functions}
\usage{
getThemes(...)

getTheme(themeId, ...)

createTheme(theme, ...)

modifyTheme(theme, ...)

deleteTheme(themeId, ...)
}
\arguments{
\item{...}{additional request parameters}

\item{themeId}{a Theme id number}

\item{theme}{a list containing Theme fields}
}
\value{
a list containing a Theme fields or a list of Themes
}
\description{
Theme Functions
}
\references{
Shopify API Theme documentation: \url{https://help.shopify.com/api/reference/online_store/theme}
}
