% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Transaction.R
\name{Transaction}
\alias{Transaction}
\alias{getTransactions}
\alias{getTransactionsCount}
\alias{getTransaction}
\alias{createTransaction}
\title{Transaction Functions}
\usage{
getTransactions(orderId, ...)

getTransactionsCount(orderId, ...)

getTransaction(orderId, transactionId, ...)

createTransaction(orderId, transaction, ...)
}
\arguments{
\item{orderId}{an \code{\link{Order}} id number}

\item{...}{additional request parameters}

\item{transactionId}{a Transaction id number}

\item{transaction}{a list containing Transaction fields}
}
\value{
a list containing a Transaction fields or a list of Transactions
}
\description{
Transaction Functions
}
\references{
Shopify API Transaction documentation: \url{https://help.shopify.com/api/reference/orders/transaction}
}
