% Generated by roxygen2 (4.0.1): do not edit by hand
\name{Webhook}
\alias{Webhook}
\alias{createWebhook}
\alias{deleteWebhook}
\alias{getWebhook}
\alias{getWebhooks}
\alias{getWebhooksCount}
\alias{modifyWebhook}
\title{Webhook Functions}
\usage{
getWebhooks(...)

getWebhooksCount(...)

getWebhook(webhookId, ...)

createWebhook(webhook, ...)

modifyWebhook(webhook, ...)

deleteWebhook(webhookId, ...)
}
\arguments{
\item{webhookId}{a Webhook id number}

\item{webhook}{a list containing Webhook fields}

\item{...}{additional request parameters}
}
\value{
a list containing a Webhook fields or a list of Webhooks
}
\description{
Webhook Functions
}
\references{
Shopify API Webhook documentation: \url{http://docs.shopify.com/api/webhook}
}

